Version 0.1 (pre-alfa)

+++ dndadmin/game/views.py
@@ -1,9 +1,13 @@
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.http import HttpResponseRedirect
-from django.views.generic import ListView, DetailView, CreateView, RedirectView, FormView
+import random
+
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.core.exceptions import PermissionDenied
+from django.http import Http404
+from django.shortcuts import redirect
+from django.views.generic import ListView, DetailView, CreateView, RedirectView, FormView, View
 
 from characterapp.models import Character
-from characterapp.utils import BaseMixin, CharacterListMixin
+from characterapp.utils import BaseMixin
 from game.forms import GameForm, GameSearchForm, HitPointForm
 from game.models import Game, Encounter, EncounterCharacter
 from game.utils import MasterInfoMixin, AddCharacterMixin
@@ -27,16 +31,19 @@ class GameListView(LoginRequiredMixin, MasterInfoMixin, ListView):
             game_objects = game_objects.filter(name__icontains=self.request.GET.get('name'))
         return game_objects
 
-class GameDetailView(LoginRequiredMixin, DetailView):
+class GameDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
     model = Game
     context_object_name = 'game'
     template_name = 'game/game_detail.html'
+
+    def test_func(self):
+        return self.request.user == self.get_object().master
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['title'] = context['game'].name
         context['characters'] = context['game'].characters.all()
-        context['encounters'] = context['game'].encounters.all()
+        context['encounters'] = context['game'].encounters.all().order_by('-time_start')
         return context
 
 class CreateGameView(LoginRequiredMixin, BaseMixin, CreateView):
@@ -51,11 +58,14 @@ class CreateGameView(LoginRequiredMixin, BaseMixin, CreateView):
         game.save()
         return super().form_valid(form)
 
-class EncounterDetailView(LoginRequiredMixin, BaseMixin, DetailView):
+class EncounterDetailView(LoginRequiredMixin, BaseMixin, UserPassesTestMixin, DetailView):
     model = Encounter
     context_object_name = 'encounter'
     template_name = 'game/encounter_detail.html'
     title_page = "Бой"
+
+    def test_func(self):
+        return self.request.user == self.get_object().game.master
 
     def get_context_data(self, **kwargs):
         context =  super().get_context_data(**kwargs)
@@ -63,20 +73,19 @@ class EncounterDetailView(LoginRequiredMixin, BaseMixin, DetailView):
         return context
 
 
-
-class CreateEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
+class CreateEncounter(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
         game = Game.objects.get(pk=self.kwargs['pk'])
-        encounter = Encounter.objects.create(game = game)
+        if not game.check_master(request.user):
+            raise PermissionDenied
+        encounter = Encounter.objects.create(game=game)
         encounter.save()
         characters = game.characters.all()
         for character in characters:
-            enc_char = EncounterCharacter.objects.create(character = character, encounter = encounter)
+            enc_char = EncounterCharacter.objects.create(character=character, encounter=encounter, hp=character.hp,
+                                                         max_hp=character.max_hp)
             enc_char.save()
-        return encounter.get_absolute_url()
+        return redirect(encounter.get_absolute_url())
 
 class CharacterAddToGameListView(LoginRequiredMixin, AddCharacterMixin, ListView):
     add_url = 'game:game-add-character-redirects'
@@ -85,51 +94,60 @@ class CharacterAddToGameListView(LoginRequiredMixin, AddCharacterMixin, ListView
         queryset = super().get_queryset()
         return queryset.filter(is_player=True).exclude(games__in=[self.kwargs['pk']])
 
-class AddCharacterToGameRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
+class AddCharacterToGame(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
         game = Game.objects.get(pk=self.kwargs['game_id'])
         character = Character.objects.get(pk=self.kwargs['character_id'])
         game.characters.add(character)
         game.save()
-        return game.get_absolute_url()
+        return redirect(game.get_absolute_url())
 
 class HeroAddToEncounterListView(LoginRequiredMixin, AddCharacterMixin, ListView):
     add_url = 'game:encounter-add-hero-redirects'
 
-class HeroAddToEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
+class HeroAddToEncounter(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
         encounter = Encounter.objects.get(pk=self.kwargs['game_id'])
-        encounter_character = EncounterCharacter.objects.create(encounter_id = self.kwargs['game_id'], character_id = self.kwargs['character_id'])
+        if not encounter.game.check_master(request.user):
+            raise PermissionDenied
+        character = Character.objects.get(pk=self.kwargs['character_id'])
+        if encounter.is_complete:
+            raise Http404
+        encounter_character = EncounterCharacter.objects.create(encounter=encounter, character=character,
+                                                                hp=character.hp, max_hp=character.max_hp)
+        if not encounter_character.character.is_player:
+            random_initiative = random.randint(1, 20)
+            encounter_character.set_initiative(random_initiative)
         encounter_character.save()
-        return encounter.get_absolute_url()
+        return redirect(encounter.get_absolute_url())
 
 class EnemyAddToEncounterListView(LoginRequiredMixin, AddCharacterMixin, ListView):
     add_url = 'game:encounter-add-enemy-redirects'
 
-class EnemyAddToEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
+class EnemyAddToEncounter(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
         encounter = Encounter.objects.get(pk=self.kwargs['game_id'])
-        encounter_character = EncounterCharacter.objects.create(encounter_id = self.kwargs['game_id'], character_id = self.kwargs['character_id'], is_enemy=True)
+        if not encounter.game.check_master(request.user):
+            raise PermissionDenied
+        character = Character.objects.get(pk=self.kwargs['character_id'])
+        random_initiative = random.randint(1, 20)
+        if encounter.is_complete:
+            raise Http404
+        encounter_character = EncounterCharacter.objects.create(encounter_id=self.kwargs['game_id'],
+                                                                character_id=self.kwargs['character_id'],
+                                                                hp=character.hp, max_hp=character.max_hp, is_enemy=True)
+        encounter_character.set_initiative(random_initiative)
         encounter_character.save()
-        return encounter.get_absolute_url()
+        return redirect(encounter.get_absolute_url())
 
-class DeleteCharacterFromEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-    def get_redirect_url(self, *args, **kwargs):
+class DeleteCharacterFromEncounter(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
         character = EncounterCharacter.objects.get(pk=self.kwargs['pk'])
+        if not character.encounter.game.check_master(request.user):
+            raise PermissionDenied
         character.delete()
         encounter = Encounter.objects.get(pk=self.kwargs['encounter_id'])
-        return encounter.get_absolute_url()
+        return redirect(encounter.get_absolute_url())
 
 class DamageFormView(LoginRequiredMixin, BaseMixin, FormView):
     template_name = 'game/game_form.html'
@@ -137,8 +155,10 @@ class DamageFormView(LoginRequiredMixin, BaseMixin, FormView):
     title_page = "Нанести урон"
 
     def form_valid(self, form):
-        character = EncounterCharacter.objects.get(pk = self.kwargs['pk']).character
-        character.get_damage(form.cleaned_data['value'])
+        character = EncounterCharacter.objects.get(pk = self.kwargs['pk'])
+        if not character.encounter.game.check_master(self.request.user):
+            raise PermissionDenied
+        character.make_damage(form.cleaned_data['value'])
         return super().form_valid(form)
 
     def get_success_url(self):
@@ -150,8 +170,10 @@ class HealthFormView(LoginRequiredMixin, BaseMixin, FormView):
     title_page = "Лечение"
 
     def form_valid(self, form):
-        character = EncounterCharacter.objects.get(pk = self.kwargs['pk']).character
-        character.get_health(form.cleaned_data['value'])
+        character = EncounterCharacter.objects.get(pk = self.kwargs['pk'])
+        if not character.encounter.game.check_master(self.request.user):
+            raise PermissionDenied
+        character.make_health(form.cleaned_data['value'])
         return super().form_valid(form)
 
     def get_success_url(self):
@@ -164,33 +186,44 @@ class SetInitiativeFormView(LoginRequiredMixin, BaseMixin, FormView):
 
     def form_valid(self, form):
         character = EncounterCharacter.objects.get(pk = self.kwargs['pk'])
+        if not character.encounter.game.check_master(self.request.user):
+            raise PermissionDenied
         character.set_initiative(form.cleaned_data['value'])
         return super().form_valid(form)
 
     def get_success_url(self):
         return EncounterCharacter.objects.get(pk = self.kwargs['pk']).encounter.get_absolute_url()
 
-class StartEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
-        encounter = Encounter.objects.get(pk = self.kwargs['pk'])
+class StartEncounterView(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
+        encounter = Encounter.objects.get(pk=self.kwargs['pk'])
+        if not encounter.game.check_master(request.user):
+            raise PermissionDenied
         if encounter.is_start:
             return encounter.get_absolute_url()
         else:
             encounter.start()
-            return encounter.get_absolute_url()
+            return redirect(encounter.get_absolute_url())
 
-class NextStepEncounterRedirect(LoginRequiredMixin, RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args, **kwargs):
-        encounter = Encounter.objects.get(pk = self.kwargs['pk'])
+class NextStepEncounter(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
+        encounter = Encounter.objects.get(pk=self.kwargs['pk'])
+        if not encounter.game.check_master(request.user):
+            raise PermissionDenied
+        if encounter.is_complete:
+            raise Http404
         encounter.next_step()
-        return encounter.get_absolute_url()
+        return redirect(encounter.get_absolute_url())
 
+class CloseEncounterView(LoginRequiredMixin, View):
+    def post(self, request, *args, **kwargs):
+        encounter = Encounter.objects.get(pk=self.kwargs['pk'])
+        if not encounter.game.check_master(request.user):
+            raise PermissionDenied
+        if encounter.is_complete:
+            raise Http404
+        encounter.close_encounter()
+        return redirect(encounter.get_absolute_url())
